feat(web): add settings to command search
